[#14] refactor : 사용자 마스킹파일 모듈화 utils/masking
사용자이름과 휴대폰을 마스킹하는 함수를 모듈화 시켰습니다.
